Added PowerControls.Domain.Configs.UnitTest.Ders and tests.

+++ paskalON.PowerControls/paskalON.PowerControls.Domain.Configs/Ders/DerUnitPowerControlConfig.cs
@@ -48,7 +48,7 @@ namespace paskalON.PowerControls.Domain.Configs.Ders
             {
                 if (value != null)
                 {
-                    ArgumentOutOfRangeException.ThrowIfNegative((ushort)value);
+                    ArgumentOutOfRangeException.ThrowIfNegative((double)value);
                 }
 
                 field = value;
